Combine `<PageCreate>` and `<PageEdit>` components
